ovn-nbd: Calculate 'up' state for logical ports.
When the state of the chassis column in the Bindings table changes for
any row, ovn-nbd will notice and trigger recalculating the 'up' state
for all logical ports.

This can be tested manually by starting up ovs-sandbox with ovn support
enabled, running ovn-nbd, and then running the following commands:

  ovn-nbctl lswitch-add sw0
  ovn-nbctl lport-add sw0-port0 sw0
  port_uuid=$(ovn-nbctl lport-list sw0 | awk '{print $1}')
  ovsdb-client transact "[\"OVN\",\
    {\"uuid-name\":\"rowd4eca046_9233_4094_bc55_e28dd49217f9\",\
    \"row\":{\"logical_port\":\"$port_uuid\",\"chassis\":\"hostname\"},\
    \"op\":\"insert\",\"table\":\"Bindings\"}]"

ovn-nbd will then see that the 'chassis' column is set in the Bindings
row for the logical port and will mark the logical port as 'up' in the
northbound db.

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Ben Pfaff <blp@nicira.com>