ovn-nbd: Populate OVN Bindings table.
When ovn-nbd gets notified that the contents of the OVN_Northbound
database have changed, it will now ensure that the contents of the
Bindings table in the OVN database are up to date.  It will create a
binding if none exists.  If one does exist, it will update its
contents as necessary, though today the only thing supported is the
list of MAC addresses.  Finally, any binding associated with a logical
port that has been deleted will be removed from the OVN database.

Signed-off-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Ben Pfaff <blp@nicira.com>