ovn-sbctl: Fix a few minor typesetting errors.
In manpages, \- is used primarily for literal "-" in literal text, that is,
it should usually be used for bold (literal) text.  (It's also used for
minus signs in mathematics.)  It should not generally be used for hyphens
in English text, as it was used here in italic (metasyntactic variable)
text.  This fixes the problem.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Russell Bryant <russell@ovn.org>