ovn-controller: Log chassis claiming lport only when changes occur.
With recent OVN commits, the logic for a chassis to claim
or release a logical port was consolidated.  This is a
good thing.  However, there was a logic change that
resulted in VLOG_INFO being generated every time on the
ovn-controller.  This patch changes the logic so that
VLOG_INFO is only generated when there is a change, for
example when the chassis claims an lport the first time.

Signed-off-by: Mickey Spiegel <mickeys.dev@gmail.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>